requirements.txt
fastapi
uvicorn
pandas
openpyxl            # 处理 Excel
python-docx         # 处理 Word
python-dotenv       # 读取 .env
langchain
langchain-google-genai  # ✅ 核心变动：适配 Google Gemini
langchain-core
pydantic